Pull the mock API object into the one module where it's used.
Eventually we might want to migrate it back out, but in the short term this
is a better arrangement because it was being used in a "common" module whereas
the mock was being defined in a server-specific module.
