+++ local-modules/testing-server/Mocks.js
@@ -2,24 +2,6 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-class FakeApiObject {
-  constructor() {
-    this.initialized = true;
-  }
-
-  static get API_NAME() {
-    return 'FakeApiObject';
-  }
-
-  toApi() {
-    return ['fake argument', 0, 1, 2];
-  }
-
-  static fromApi(arguments_unused) {
-    return new FakeApiObject();
-  }
-}
-
 /**
  * A collection of methods for creating mock objects needed for Bayou unit
  * testing.
@@ -32,9 +14,5 @@ export default class Mocks {
                      followRedirects = true,
                      maxRedirects = 10) {
     return { uri, method, headers, timeout, followRedirects, maxRedirects };
-  }
-
-  static apiObject() {
-    return new FakeApiObject();
   }
 }
